Move RPC handlers to internal/server/apiserver package
Extract all Connect RPC handler methods into a dedicated apiserver
package. The internal/server package becomes a thin orchestration
layer that wires together auth, routing, and sub-servers.
