+++ internal/server/apiserver/routing_rules_test.go
@@ -1,4 +1,4 @@
-package server
+package apiserver
 
 import (
 	"testing"
