+++ internal/server/apiserver/runner_test.go
@@ -1,4 +1,4 @@
-package server
+package apiserver
 
 import (
 	"testing"
